Created admin-tokens
Run this program to generate random voter ID tokens. The election
authority should save these tokens; whenever a vote comes in, it should
check whether its token is one of the saved tokens. This program is
designed so that the user can redirect stdout to a file to capture the
tokens, one on each line.