Added check of number of votes to tally-votes
Mathematically, it is possible to send a single ballot with one of the
following irregularities:
* At least one candidate has more than one vote
* More than one candidate has a vote
* Both of the above

There is a simple way to check for this, though: the total number of
votes for all candidates should equal the number of ballots received.
This check is now performed by `tally-votes` at the very end. If the two
numbers do not match, the election results cannot be trusted.