Added reverse ballot checking
`cast-vote` now includes a reverse ballot. If `tally-votes` processes
the reverse ballot like it does the regular ballot, then the result is
the same as the regular result, except that the candidates are reversed.
If the results from the reverse ballots are un-reversed, they should
match the results from the regular ballots. If they do not match, the
election results cannot be trusted.