Added detection of extra candidates
Voters should not be allowed to introduce new candidates into the
election because they will not be seen by tally-votes. Now, tally-votes
will check that there are no more candidates than were specified. It
does this by making sure that the sum of all votes is less than or equal
to the sum if all voters voted for the last candidate. It also checks
this for the reverse sum.